fixed some issues

+++ hwy/contrib/random/random.h
@@ -180,9 +180,9 @@ class VectorXoshiro {
 
   HWY_INLINE VU64 operator()() noexcept { return Next(); }
 
-  AlignedVector<std::size_t> operator()(const std::uint64_t n) {
-    AlignedVector<std::size_t> result(n);
-    const ScalableTag<std::size_t> tag{};
+  AlignedVector<std::uint64_t> operator()(const std::size_t n) {
+    AlignedVector<std::uint64_t> result(n);
+    const ScalableTag<std::uint64_t> tag{};
     auto s0 = Load(tag, state_[{0}].data());
     auto s1 = Load(tag, state_[{1}].data());
     auto s2 = Load(tag, state_[{2}].data());
@@ -230,7 +230,7 @@ class VectorXoshiro {
     return Mul(real, MUL_VALUE);
   };
 
-  AlignedVector<double> Uniform(const std::uint64_t n) {
+  AlignedVector<double> Uniform(const std::size_t n) {
     AlignedVector<double> result(n);
     const ScalableTag<std::size_t> tag{};
     const ScalableTag<double> real_tag{};
@@ -285,7 +285,7 @@ class VectorXoshiro {
 
  private:
   StateType state_;
-  const std::size_t streams;
+  const std::uint64_t streams;
 
   HWY_INLINE static VU64 Update(VU64 &s0, VU64 &s1, VU64 &s2,
                                 VU64 &s3) noexcept {
